# frontend/openapi/fixture.yaml
openapi: 3.1.0
info:
  title: Amezo API (fixture)
  version: 0.0.0
  description: >
    Placeholder spec so the codegen pipeline (springdoc -> OpenAPI ->
    openapi-typescript) has something to run against before the Spring
    backend exposes its own. Modeled on docs/api-design.md's documented
    GET /products contract, not the backend's current partial
    implementation (q-only as of this writing — see ProductController.java).
    Swap `gen:api` in package.json to point at the live springdoc endpoint
    once the backend catches up to this contract; this file goes away then.
paths:
...
  /orders:
    post:
      operationId: checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /sellers/me/orders/{orderId}:
    get:
      operationId: sellerGetOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerOrderDetail'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /api/v1/orders:
    get:
      operationId: listBuyerOrders
      description: >-
        The signed-in buyer's own orders, newest first. One entry per seller
        order, so a checkout that split across two sellers shows two rows.
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/OrderStatus' }
        - name: group
          in: query
          description: Coarse filter the My Orders tabs use, applied on top of status.
          schema: { type: string, enum: [all, in_progress, delivered, refunds] }
        - name: q
          in: query
          description: Matches order reference and product titles.
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerOrderSummaryPage'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/orders/{orderId}:
    get:
      operationId: getBuyerOrder
      description: 404 (not 403) when the order exists but is not the caller's.
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyerOrderDetail'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /api/v1/sellers/me/store:
    get:
      operationId: getMyStore
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreProfile'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    patch:
      operationId: updateMyStore
      description: >-
        Partial update; an omitted field is left as it is. 409 when the handle
        is already taken, so the seller learns which field collided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreProfile'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreProfile'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /api/v1/sellers/me/store/images:
    post:
      operationId: createStoreImageUploadUrl
      description: >-
        A presigned slot for the storefront cover or logo, mirroring the
        product-image flow (POST for a URL, PUT the bytes to it, then confirm).
        Store images had no upload path at all, so coverUrl and logoUrl were
        writable on the profile with no way to produce a URL to write.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              # fileSizeBytes is required, not informational: the server signs it
              # into the presigned URL as Content-Length, so a PUT of any other
              # size fails the signature rather than quietly costing storage, and
              # it is the number the 5 MB per-file cap is checked against before
              # any URL is issued. An upload with no declared size cannot be
              # capped at all, so there is no useful behaviour behind omitting it.
              required: [slot, contentType, fileSizeBytes]
              properties:
                slot: { type: string, enum: [COVER, LOGO] }
                contentType: { type: string }
                fileSizeBytes: { type: integer }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [id, uploadUrl, expiresAt]
                properties:
                  id: { type: string, format: uuid }
                  uploadUrl: { type: string }
                  expiresAt: { type: string, format: date-time }
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/sellers/me/store/images/confirm:
    post:
      operationId: confirmStoreImage
      description: >-
        Promotes an uploaded slot to the live storefront and returns the updated
        profile, so the caller does not have to re-read it to learn the URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, slot]
              properties:
                id: { type: string, format: uuid }
                slot: { type: string, enum: [COVER, LOGO] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreProfile'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /api/v1/sellers/me/orders/facets:
    get:
      operationId: getSellerOrderFacets
      description: >-
        Counts and bucket values for the order tabs. A separate resource rather
        than a field on the list, because the tabs show every bucket at once:
        embedding them in a filtered response would zero every tab but the one
        being viewed. Honours q and the date range, and deliberately ignores
        status - that is what the tabs are choosing between.
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacetList'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /api/v1/orders/facets:
    get:
      operationId: getBuyerOrderFacets
      description: >-
        Counts for the buyer's order tabs. Keyed by the same group values GET
        /api/v1/orders takes, not by raw status - the tabs are groups.
      parameters:
        - name: q
          in: query
          schema: { type: string }
        # The same window the list takes, deliberately. A separate `period`
        # vocabulary here meant the client and the server each decided what
        # "past 3 months" was, so a tab could count a window the list it opened
        # did not use. Dates are computed once, on the client, from the select.
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacetList'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
  /api/v1/sellers/me/orders/{orderId}:
    get:
      operationId: sellerGetOrderV1
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerOrderRowDetail'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    patch:
      operationId: sellerUpdateOrder
      description: >-
        Advances the order. Replaces POST /sellers/me/orders/{orderId}/ship,
        which put the verb in the path. PLACED to PACKED takes parcels;
        PACKED/PLACED to SHIPPED takes a tracking number. Any other transition
        is 409.
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSellerOrder'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerOrderRowDetail'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  schemas:
...
    Address:
      type: object
      required: [fullName, line1, city, state, postalCode, country]
      properties:
        fullName: { type: string }
        line1: { type: string }
        # The one optional field, matching CheckoutAddressRequest's validation
        # exactly - it is not marked required on the form either.
        line2: { type: string, nullable: true }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        # ISO 3166-1 alpha-2, the same code GET /countries serves and
        # @ValidCountryCode enforces - never a display name.
        country: { type: string, minLength: 2, maxLength: 2 }
...
    OrderLineDetail:
      type: object
      required: [id, productTitle, variantLabel, quantity, unitPrice]
      properties:
        id: { type: string, format: uuid }
        productTitle: { type: string }
        variantLabel: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        # Optional: the real backend's shared OrderLineResponse record (also
        # used by checkout's confirmation response) always includes this,
        # but the seller order-detail UI doesn't need it - total is shown
        # per-order, not per-line.
        lineTotal: { type: number }
    CheckoutAddress:
      type: object
      required: [fullName, line1, city, state, postalCode, country]
      properties:
        fullName: { type: string }
        line1: { type: string }
        line2: { type: string, nullable: true }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        # An ISO 3166-1 alpha-2 code from GET /countries. Validated on the server, so
        # a client that skips the selector cannot store "XX" or the plausible-but-
        # wrong "UK".
        country: { type: string, minLength: 2, maxLength: 2 }
    CheckoutLine:
      type: object
      required: [variantId, quantity]
      properties:
        variantId: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1 }
        expectedUnitPrice: { type: number, nullable: true }
    CheckoutRequest:
      type: object
      required: [email, phone, shippingAddress, sameAsShipping, lines]
      properties:
        email: { type: string, format: email }
        phone: { type: string }
        shippingAddress: { $ref: '#/components/schemas/CheckoutAddress' }
        sameAsShipping: { type: boolean }
        billingAddress:
          allOf: [{ $ref: '#/components/schemas/CheckoutAddress' }]
          nullable: true
        lines:
          type: array
          items: { $ref: '#/components/schemas/CheckoutLine' }
    OrderLineResponse:
      type: object
      required: [id, productTitle, variantLabel, quantity, unitPrice, lineTotal]
      properties:
        id: { type: string, format: uuid }
        productTitle: { type: string }
        variantLabel: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        lineTotal: { type: number }
    OrderResponse:
      type: object
      required: [id, placedAt, lines, total]
      properties:
        id: { type: string, format: uuid }
        placedAt: { type: string, format: date-time }
        lines:
          type: array
          items: { $ref: '#/components/schemas/OrderLineResponse' }
        total: { type: number }
    SellerOrderDetail:
      type: object
      required: [id, buyerEmail, placedAt, total, status, lines]
      properties:
        id: { type: string, format: uuid }
        buyerEmail: { type: string }
        placedAt: { type: string, format: date-time }
        total: { type: number }
        status: { $ref: '#/components/schemas/OrderStatus' }
        trackingNumber: { type: string, nullable: true }
        shippedAt: { type: string, format: date-time, nullable: true }
        lines:
          type: array
          items: { $ref: '#/components/schemas/OrderLineDetail' }
...
    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }

    # --- /api/v1 shapes -------------------------------------------------------

    # Supersedes the three-value status on the unversioned seller endpoints.
    # PACKED is the handover step the seller screens add; IN_TRANSIT and
    # OUT_FOR_DELIVERY are what the buyer's tracking timeline shows once a
    # carrier is scanning.
    OrderStatus:
      type: string
      description: >-
        Read-only. PLACED through DELIVERED is the fulfilment lifecycle the
        seller drives; CANCELLED and REFUNDED are outcomes the server decides.

        REFUNDED is DERIVED, never written: the server reports it when the
        order's refund request reaches REFUNDED. Refunds are modelled once, by
        refund_request, with their own state machine, approval and return steps
        and money. An order status a client could also write would be a second
        source of truth that nothing reconciles the first time somebody settles
        a refund through the queue instead. See SellerOrderTransition for what a
        seller may actually write.
      enum:
        [PLACED, PACKED, SHIPPED, IN_TRANSIT, OUT_FOR_DELIVERY, DELIVERED, CANCELLED, REFUNDED]
    SellerOrderTransition:
      type: string
      description: >-
        The statuses a seller may write, which is deliberately narrower than
        OrderStatus. The three carrier states are system-driven, and REFUNDED is
        derived from the refund request rather than declared here.
      enum: [PACKED, SHIPPED, CANCELLED]
...
    PaymentSummary:
      type: object
      required: [brand, last4]
      properties:
        brand: { type: string }
        last4: { type: string, minLength: 4, maxLength: 4 }
    OrderTimelineEntry:
      type: object
      required: [code, label, completed]
      properties:
        code: { type: string, enum: [PLACED, PACKED, SHIPPED, IN_TRANSIT, OUT_FOR_DELIVERY, DELIVERED] }
        label: { type: string }
        at: { type: string, format: date-time, nullable: true }
        # True while `at` is a projection rather than something that happened.
        estimated: { type: boolean, default: false }
        completed: { type: boolean }
        detail: { type: string, nullable: true }
    ShipmentInfo:
      type: object
      properties:
        carrier: { type: string, nullable: true }
        trackingNumber: { type: string, nullable: true }
        trackingUrl: { type: string, nullable: true }
        shippedAt: { type: string, format: date-time, nullable: true }
        estimatedDeliveryAt: { type: string, format: date-time, nullable: true }
        deliveredAt: { type: string, format: date-time, nullable: true }
        deliveryNote: { type: string, nullable: true }
    BuyerOrderLine:
      type: object
      required: [id, productRef, productTitle, variantLabel, quantity, unitPrice, lineTotal]
      properties:
        id: { type: string, format: uuid }
        productRef: { type: string }
        productTitle: { type: string }
        variantLabel: { type: string }
        thumbnailUrl: { type: string, nullable: true }
        quantity: { type: integer }
        unitPrice: { type: number }
        lineTotal: { type: number }
        # Set when this line is inside an open refund request.
        refundRequestId: { type: string, format: uuid, nullable: true }
        refundStatus:
          oneOf:
            - $ref: '#/components/schemas/RefundStatus'
          nullable: true
    StoreRef:
      type: object
      required: [name, handle]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        handle: { $ref: '#/components/schemas/StoreHandle' }
    BuyerOrderSummary:
      type: object
      required: [id, reference, placedAt, status, total, currency, itemCount, seller]
      properties:
        id: { type: string, format: uuid }
        # Short human reference the screens print; the uuid stays the key.
        reference: { type: string }
        placedAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/OrderStatus' }
        total: { type: number }
        currency: { type: string }
        itemCount: { type: integer }
        seller: { $ref: '#/components/schemas/StoreRef' }
        previewLines:
          type: array
          items: { $ref: '#/components/schemas/BuyerOrderLine' }
        shipment: { $ref: '#/components/schemas/ShipmentInfo' }
        openRefundRequestId: { type: string, format: uuid, nullable: true }
        # The list needs the status, not just the id: the card prints a badge
        # for it, and a refund that has been approved or declined does not read
        # the same as one nobody has looked at yet.
        openRefundStatus:
          oneOf:
            - $ref: '#/components/schemas/RefundStatus'
          nullable: true
    BuyerOrderSummaryPage:
      type: object
      required: [content, page, totalElements, totalPages]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/BuyerOrderSummary' }
        page: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
    BuyerOrderDetail:
      type: object
      required:
        [id, reference, placedAt, status, seller, lines, subtotal, shipping, tax, total, currency, timeline, shippingAddress]
      properties:
        id: { type: string, format: uuid }
        reference: { type: string }
        placedAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/OrderStatus' }
        seller: { $ref: '#/components/schemas/StoreRef' }
        lines:
          type: array
          items: { $ref: '#/components/schemas/BuyerOrderLine' }
        subtotal: { type: number }
        shipping: { type: number }
        tax: { type: number }
        total: { type: number }
        currency: { type: string }
        timeline:
          type: array
          items: { $ref: '#/components/schemas/OrderTimelineEntry' }
        shipment: { $ref: '#/components/schemas/ShipmentInfo' }
        shippingAddress: { $ref: '#/components/schemas/Address' }
        billingAddress:
          oneOf:
            - $ref: '#/components/schemas/Address'
          nullable: true
        payment:
          oneOf:
            - $ref: '#/components/schemas/PaymentSummary'
          nullable: true
        refundRequests:
          type: array
          items: { $ref: '#/components/schemas/RefundRequestSummary' }
        # Server-owned, so no screen re-derives the return window from dates.
        canRequestRefund: { type: boolean }
        refundWindowEndsAt: { type: string, format: date-time, nullable: true }
...
    StoreStatus:
      type: string
      enum: [OPEN, VACATION, CLOSED]
    StoreHandle:
      type: string
      description: >-
        The store's URL segment (amezo.com/stores/{handle}). Lowercase letters,
        digits and inner dashes only - never a leading or trailing dash, so the
        printed URL never ends in one. Case-insensitive on lookup and unique
        across stores; PATCH /api/v1/sellers/me/store answers 409 when taken.
      pattern: '^[a-z0-9][a-z0-9-]{0,37}[a-z0-9]$'
      minLength: 2
      maxLength: 39
      example: northwind-supply
    StoreProfile:
      type: object
      required: [id, name, handle, status]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        handle: { $ref: '#/components/schemas/StoreHandle' }
        tagline: { type: string, nullable: true }
        location: { type: string, nullable: true }
        foundedYear: { type: integer, nullable: true }
        supportEmail: { type: string, nullable: true }
        about: { type: string, nullable: true }
        coverUrl: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/StoreStatus' }
        vacationNote: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
    UpdateStoreProfile:
      type: object
      properties:
        name: { type: string }
        handle: { $ref: '#/components/schemas/StoreHandle' }
        tagline: { type: string }
        location: { type: string }
        # Nullable, because "Selling since" is a box the seller can empty and the
        # storefront preview then reads "New seller". An omitted field is left as
        # it is and a blank string clears the text fields, but an integer has no
        # blank - so null is the only way to say "remove the founding year", and
        # without it a year could be set and never taken back out.
        foundedYear: { type: integer, nullable: true }
        supportEmail: { type: string }
        about: { type: string }
        coverUrl: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/StoreStatus' }
        vacationNote: { type: string }
...
    FacetList:
      type: object
      required: [facets]
      properties:
        facets:
          type: array
          items: { $ref: '#/components/schemas/Facet' }
...
    SellerOrderRowDetail:
      type: object
      required:
        [id, reference, buyerEmail, recipientName, placedAt, status, lines, subtotal, shipping, total, currency, shippingAddress]
      properties:
        id: { type: string, format: uuid }
        reference: { type: string }
        buyerEmail: { type: string }
        recipientName: { type: string }
        placedAt: { type: string, format: date-time }
        status: { $ref: '#/components/schemas/OrderStatus' }
        lines:
          type: array
          items: { $ref: '#/components/schemas/SellerOrderLine' }
        subtotal: { type: number }
        shipping: { type: number }
        tax: { type: number }
        total: { type: number }
        currency: { type: string }
        shippingAddress: { $ref: '#/components/schemas/Address' }
        shipment: { $ref: '#/components/schemas/ShipmentInfo' }
        packedAt: { type: string, format: date-time, nullable: true }
        parcels: { type: integer, nullable: true }
        refundRequests:
          type: array
          items: { $ref: '#/components/schemas/RefundRequestSummary' }
...
    UpdateSellerOrder:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/SellerOrderTransition' }
        # PLACED -> PACKED
        parcels: { type: integer, minimum: 1 }
        packedBy: { type: string }
        # PACKED -> SHIPPED. The tracking number is issued by the platform on
        # handover, not typed by the seller, so it is not writable here.
        handoverMethod: { $ref: '#/components/schemas/HandoverMethod' }
        hub: { type: string }
        # Shown to the buyer on the order's timeline.
        note: { type: string }
        # When it actually happened, for a seller recording yesterday's handover
        # today. Omitted means now. The server refuses a future one.
        occurredAt: { type: string, format: date-time }
...